### 2021-07-26 #455 @JeffreyRStevens
excluder: checks for exclusion criteria in online data
Submitting Author Name: Jeffrey Stevens
Submitting Author Github Handle: @JeffreyRStevens
Repository:  https://github.com/JeffreyRStevens/excluder
Version submitted: 0.2.2
Submission type: Standard
Version accepted: TBD

---

```
Package: excluder
Title: Checks for Exclusion Criteria in Online Data
Version: 0.2.2
    person(given = "Jeffrey R.",
           family = "Stevens",
           email = "jeffrey.r.stevens@gmail.com",
Description: Data that are collected through online sources such as Mechanical
            Turk may require excluding data because of IP address duplication,
            geolocation, or completion duration. This package facilitates
            exclusion of these data for Qualtrics datasets.
    iptools,
    janitor,
    maps,
    rlang
    covr,
    lifecycle
```

The package falls under data munging because it processes data from Qualtrics or other online sources by checking for, marking, and excluding rows of data frames for common exclusion criteria (e.g., IP addresses outside of the United States or duplicate entries from the same location/IP address).

The target audience is data scientists using Qualtrics or other online systems to collect data from participants (e.g., Mechanical Turk workers). Ensuring good data quality from these participants can be tricky. For instance, while Mechanical Turk in theory screens workers based on location (e.g., if you want to restrict your participant pool to workers in the United States), this is not necessarily represented in the data. Finding the tools to screen for IP address location can be tricky, and this package simplifies checking for and excluding participants based on common data that Qualtrics reports such as geolocation, IP address, duplicate records from the same location, participant screen resolution, participant progress through the survey, and survey completion duration.

There are no similar packages to my knowledge. The {qualtRics} package at rOpenSci focuses on importing data from Qualtrics. The {MTurkR} package directly interfaces with the MTurk Requestor API, but the APIs have been deprecated and the package has been removed from CRAN.

Yes, it seems to comply with this guidance. Depending on the data that the user collects, there could be personally identifiable information accessed by this package. In particular, IP addresses that are recorded by Qualtrics can be processed with this package. Note that the package only works with personally identifiable information from data sets that already exist on the users' local file system, and the package does not collect or transmit data in any way. The package also includes a function `deidentify()` that the user can use to strip location, IP address, language and even participant computer information (e.g., operating system, web browser, screen resolution) from the data frames to deidentify them.

https://github.com/ropensci/software-review/issues/454
