### 2021-07-30 #455 @JeffreyRStevens
The only check that failed was "Package does not have a 'contributing.md' file". Does this mean that the package should **not** have a contributing.md file? So should I just remove the file and remove all references to Contributing to this package?
