### 2021-07-30 #455 @noamross
Thanks for the submission @JeffreyRStevens! Sorry for a bit of a delay as we were working out our new automated diagnostics bot, which your review is the first to use, and you just found a bug in! Your `CONTRIBUTING.md` file is fine, we are just failing to check the right subfolder.  We'll move ahead with this :)
