### 2021-08-07 #455 @maurolepore
@JeffreyRStevens, it's my pleasure to be the handling editor of your submission.

### Editor checks:

- [x] **Automated tests:** Package has a testing suite and is tested via a CI service.

---

#### Editor comments

Congratulations! The bot and I are very happy to see the package meets rOpenSci's guidelines :+1:. I'll start looking for reviewers.

Note the following minor issues. You might want to consider before the review:

* `devtools::check()` shows:

```r
❯ checking Rd cross-references ... NOTE
  Packages unavailable to check Rd xrefs: ‘qualtRics’, ‘rgeolocate’
```

* `check_deduplicate()` seems to have a relatively complex logic. It may be justified by the goal of the function (to check for various conditions) but might be worth ensuring there is no more logic than necessary.

* Some `if()` statements seem fragile as they could get inputs of length greater than 1. Consider the argument `dupl_ip` to `check_duplicates()`:

``` r
# Good
dupl_ip <- TRUE
if (identical(dupl_ip, TRUE)) {
  message("Do something.")
}
#> Do something.

# Good
dupl_ip <- TRUE
stopifnot(length(dupl_ip) == 1L)
if (dupl_ip) {
  message("Do something.")
}
#> Do something.

# Fragile
dupl_ip <- c(TRUE, FALSE)  # This might be accidentally non-atomic
if (dupl_ip == TRUE) {
  message("Do something.")
}
#> Warning in if (dupl_ip == TRUE) {: the condition has length > 1 and only the
#> first element will be used
#> Do something.
```

* Some lines are longer than 80 characters. Wrapping at 80 characters makes the code easier to review, without having to scroll right when navigating the script.

* I see two methods to avoid the R CMD check NOTE about undefined global variables: `var <- NULL` ([e.g.](https://github.com/JeffreyRStevens/excluder/blob/1d8446c90ca467b8257592230a7b0abc8f348c48/R/collapse_exclusions.R#L45)) and [`.data$var`](https://github.com/JeffreyRStevens/excluder/blob/1d8446c90ca467b8257592230a7b0abc8f348c48/R/collapse_exclusions.R#L55). Maybe best to use only one method?

* In R/ I see a subdirectory (R/deprecated). I find this strange. How is it used?

* `devtools::test()` shows:

```r
check-mark-exclude:
Note: Using an external vector in selections is ambiguous.
i Use `all_of(location_col)` instead of `location_col` to silence this message.
i See <https://tidyselect.r-lib.org/reference/faq-external-vector.html>.
```

* The structure of the tests impairs the use of the convenient workflow with `usethis::use_r()` and `usethis::use_test()`. Consider structuring tests so that for each exported `fun()` in R/fun.R you have a test in tests/testthat/test-fun()`.

---

Reviewers: TBC
Due date:
