### 2021-08-08 #455 @maurolepore
@JeffreyRStevens, could you suggest two or three potential reviewers? Although I wouldn't pick more than one, your list will inform the type of expertise you think would be useful when reviewing {excluder}.

I'll also use other criteria as described in [How to look for reviewers](https://devdevguide.netlify.app/editorguide.html?#look-for-and-assign-reviewers).
