### 2021-08-08 #455 @JeffreyRStevens
@maurolepore thank you for your speedy initial review of {excluder}. I have pushed some changes to address your comments.
* In running `usethis::check()`, I am not able to replicate your error regarding `Packages unavailable to check Rd xrefs: ‘qualtRics’, ‘rgeolocate’`. Would you mind checking if this is still an issue in the new version? If so, does this mean that you need to have the packages installed on your local machine to pass the check?
* I'm not sure if it is possible to simplify the logic of `check_duplicates()`. I'm open to ideas from you or the reviewers, though.
* Regarding the fragile `if()` statements, are you suggesting that I check if the `dupl_id` argument is identical to TRUE or of length 1. And, if so, I assume you'd like me to look at other `if()` statements to see if the same issue applies? Just want to clarify the action items for this issue.
* I trimmed lines longer than 80 characters. The remaining lines > 80 characters are primarily messages/warnings or tests. Should I break those across multiple lines? Is the aim to reduce the number of lines > 80 or minimize the number of characters > 80? For example, in `remove_label_rows.R` lines [53-59](https://github.com/JeffreyRStevens/excluder/blob/d99bc23850c3f6f7ae3001cf0b94324b9e0a827b/R/remove_label_rows.R#L53), I broke this over multiple lines to minimize characters > 80, but it resulted  in severa lines slightly larger that 80 characters rather than just line much larger. Which is preferred?
* Regarding undefined global variables, I used `exclusions <- NULL` in `collapse_exclusions.R` because I can't figure out how to use `.data` in a way to avoid the R CMD note. If I use `.data$exclusions` to replace `exclusions` in the `unite()` function (line [55](https://github.com/JeffreyRStevens/excluder/blob/d99bc23850c3f6f7ae3001cf0b94324b9e0a827b/R/collapse_exclusions.R#L55)) or the column definition in the `mutate()` function (line [60](https://github.com/JeffreyRStevens/excluder/blob/d99bc23850c3f6f7ae3001cf0b94324b9e0a827b/R/collapse_exclusions.R#L60)), I receive a compiling error when running the check. The current version has removed the `exclusions <- NULL`, which results in the R CMD note. One option would be to reinsert the `NULL` definition and remove all `.data` from this particular function. But would I need to replace all `.data` with `NULL` definitions in all functions or just this one?
* LOL--the directory _R/deprecated_ was supposed to be added to the `.gitignore` file. It has now been removed.
* I have addressed the `all_of(location_col)` warning in the test.
* I have restructured the tests to have a test for each function.
* I have added the rOpenSci peer review badge to the README.
* Two potential reviewers would be the authors of {qualtRics}: Julia Silge and Jasper Ginn.

Thank you again for your careful review, and let me know if I missed something.
