### 2021-08-10 #455 @maurolepore
Thanks @JeffreyRStevens for responding quickly.

* Yes, I still see that R CMD NOTE and two more. FYI I updated all my packages then run `devtools::check()`. Here is a reprex -- collapsed to save space.

<summary>
Output
</summary>

``` r
devtools::load_all()
#> ℹ Loading excluder
packageVersion("excluder")
#> [1] '0.2.2'
gert::git_log(max = 1)
#> # A tibble: 1 × 6
#>   commit         author          time                files merge message
#> * <chr>          <chr>           <dttm>              <int> <lgl> <chr>
#> 1 d99bc23850c3f… Jeffrey R. Ste… 2021-08-08 15:06:45    29 FALSE "Change all te…

devtools::check()
#> ℹ Updating excluder documentation
#> ℹ Loading excluder
#> Warning: [/home/mauro/git/excluder/R/check_duplicates.R:9] @details Link
#> to unavailable package: qualtRics::fetch_survey. there is no package called
#> 'qualtRics'
#> Warning: [/home/mauro/git/excluder/R/check_duration.R:9] @details Link to
#> unavailable package: qualtRics::fetch_survey. there is no package called
#> 'qualtRics'
#> Warning: [/home/mauro/git/excluder/R/check_ip.R:9] @details Link to unavailable
#> package: qualtRics::fetch_survey. there is no package called 'qualtRics'
#> Warning: [/home/mauro/git/excluder/R/check_location.R:9] @details Link to
#> unavailable package: qualtRics::fetch_survey. there is no package called
#> 'qualtRics'
#> Warning: [/home/mauro/git/excluder/R/check_preview.R:9] @details Link to
#> unavailable package: qualtRics::fetch_survey. there is no package called
#> 'qualtRics'
#> Warning: [/home/mauro/git/excluder/R/check_progress.R:9] @details Link to
#> unavailable package: qualtRics::fetch_survey. there is no package called
#> 'qualtRics'
#> Warning: [/home/mauro/git/excluder/R/check_resolution.R:10] @details Link
#> to unavailable package: qualtRics::fetch_survey. there is no package called
#> 'qualtRics'
#> Warning: [/home/mauro/git/excluder/R/qualtrics_numeric.R:3] @description Link
#> to unavailable package: rgeolocate::ip2location. there is no package called
#> 'rgeolocate'
#> Warning: [/home/mauro/git/excluder/R/qualtrics_raw.R:3] @description Link
#> to unavailable package: rgeolocate::ip2location. there is no package called
#> 'rgeolocate'
#> Warning: [/home/mauro/git/excluder/R/qualtrics_text.R:3] @description Link
#> to unavailable package: rgeolocate::ip2location. there is no package called
#> 'rgeolocate'
#> Warning: [/home/mauro/git/excluder/R/remove_label_rows.R:7] @details Link
#> to unavailable package: qualtRics::fetch_survey. there is no package called
#> 'qualtRics'
#> Writing NAMESPACE
#> Writing NAMESPACE
#> ── Building ──────────────────────────────────────────────────────── excluder ──
#> Setting env vars:
#> • CFLAGS    : -Wall -pedantic
#> • CXXFLAGS  : -Wall -pedantic
#> • CXX11FLAGS: -Wall -pedantic
#> ────────────────────────────────────────────────────────────────────────────────
#>      checking for file ‘/home/mauro/git/excluder/DESCRIPTION’ ...  ✓  checking for file ‘/home/mauro/git/excluder/DESCRIPTION’
#>   ─  preparing ‘excluder’:
#>        checking DESCRIPTION meta-information ...  ✓  checking DESCRIPTION meta-information
#>   ─  installing the package to build vignettes
#>      creating vignettes ...  ✓  creating vignettes (2.6s)
#>   ─  checking for LF line-endings in source and make files and shell scripts
#>   ─  checking for empty or unneeded directories
#>   ─  building ‘excluder_0.2.2.tar.gz’
#>
#> ── Checking ──────────────────────────────────────────────────────── excluder ──
#> Setting env vars:
#> • _R_CHECK_CRAN_INCOMING_USE_ASPELL_: TRUE
#> • _R_CHECK_CRAN_INCOMING_REMOTE_    : FALSE
#> • _R_CHECK_CRAN_INCOMING_           : FALSE
#> • _R_CHECK_FORCE_SUGGESTS_          : FALSE
#> • NOT_CRAN                          : true
#> ── R CMD check ─────────────────────────────────────────────────────────────────
#> * using log directory ‘/tmp/Rtmpcusbks/excluder.Rcheck’
#> * using R version 4.1.0 (2021-05-18)
#> * using platform: x86_64-pc-linux-gnu (64-bit)
#> * using session charset: UTF-8
#> * using options ‘--no-manual --as-cran’
#> * checking for file ‘excluder/DESCRIPTION’ ... OK
#> * this is package ‘excluder’ version ‘0.2.2’
#> * package encoding: UTF-8
#> * checking package namespace information ... OK
#> * checking package dependencies ... OK
#> * checking if this is a source package ... OK
#> * checking if there is a namespace ... OK
#> * checking for executable files ... OK
#> * checking for hidden files and directories ... OK
#> * checking for portable file names ... OK
#> * checking for sufficient/correct file permissions ... OK
#> * checking whether package ‘excluder’ can be installed ... OK
#> * checking installed package size ... OK
#> * checking package directory ... OK
#> * checking for future file timestamps ... OK
#> * checking ‘build’ directory ... OK
#> * checking DESCRIPTION meta-information ... OK
#> * checking top-level files ... NOTE
#> Non-standard files/directories found at top level:
#>   ‘mid-guppy_reprex.R’ ‘mid-guppy_reprex.md’ ‘ok-coqui_reprex.R’
#>   ‘ok-coqui_reprex.spin.R’ ‘ok-coqui_reprex.spin.Rmd’
#> * checking for left-over files ... OK
#> * checking index information ... OK
#> * checking package subdirectories ... OK
#> * checking R files for non-ASCII characters ... OK
#> * checking R files for syntax errors ... OK
#> * checking whether the package can be loaded ... OK
#> * checking whether the package can be loaded with stated dependencies ... OK
#> * checking whether the package can be unloaded cleanly ... OK
#> * checking whether the namespace can be loaded with stated dependencies ... OK
#> * checking whether the namespace can be unloaded cleanly ... OK
#> * checking loading without being on the library search path ... OK
#> * checking dependencies in R code ... OK
#> * checking S3 generic/method consistency ... OK
#> * checking replacement functions ... OK
#> * checking foreign function calls ... OK
#> * checking R code for possible problems ... NOTE
#> collapse_exclusions: no visible binding for global variable
#>   ‘exclusions’
#> Undefined global functions or variables:
#>   exclusions
#> * checking Rd files ... OK
#> * checking Rd metadata ... OK
#> * checking Rd line widths ... OK
#> * checking Rd cross-references ... NOTE
#> Packages unavailable to check Rd xrefs: ‘qualtRics’, ‘rgeolocate’
#> * checking for missing documentation entries ... OK
#> * checking for code/documentation mismatches ... OK
#> * checking Rd \usage sections ... OK
#> * checking Rd contents ... OK
#> * checking for unstated dependencies in examples ... OK
#> * checking contents of ‘data’ directory ... OK
#> * checking data for non-ASCII characters ... OK
#> * checking LazyData ... OK
#> * checking data for ASCII and uncompressed saves ... OK
#> * checking installed files from ‘inst/doc’ ... OK
#> * checking files in ‘vignettes’ ... OK
#> * checking examples ... OK
#> * checking for unstated dependencies in ‘tests’ ... OK
#> * checking tests ...
#>   Running ‘testthat.R’
#>  OK
#> * checking for unstated dependencies in vignettes ... OK
#> * checking package vignettes in ‘inst/doc’ ... OK
#> * checking re-building of vignette outputs ... OK
#> * checking for non-standard things in the check directory ... OK
#> * checking for detritus in the temp directory ... OK
#> * DONE
#>
#> Status: 3 NOTEs
#> See
#>   ‘/tmp/Rtmpcusbks/excluder.Rcheck/00check.log’
#> for details.
#> ── R CMD check results ───────────────────────────────────── excluder 0.2.2 ────
#> Duration: 56.5s
#>
#> > checking top-level files ... NOTE
#>   Non-standard files/directories found at top level:
#>     ‘mid-guppy_reprex.R’ ‘mid-guppy_reprex.md’ ‘ok-coqui_reprex.R’
#>     ‘ok-coqui_reprex.spin.R’ ‘ok-coqui_reprex.spin.Rmd’
#>
#> > checking R code for possible problems ... NOTE
#>   collapse_exclusions: no visible binding for global variable
#>     ‘exclusions’
#>   Undefined global functions or variables:
#>     exclusions
#>
#> > checking Rd cross-references ... NOTE
#>   Packages unavailable to check Rd xrefs: ‘qualtRics’, ‘rgeolocate’
#>
#> 0 errors ✓ | 0 warnings ✓ | 3 notes x
```

<summary>
Session info
</summary>

``` r
sessioninfo::session_info()
#> ─ Session info ───────────────────────────────────────────────────────────────
#>  setting  value
#>  version  R version 4.1.0 (2021-05-18)
#>  os       Ubuntu 20.04.2 LTS
#>  system   x86_64, linux-gnu
#>  ui       X11
#>  language (EN)
#>  collate  en_US.UTF-8
#>  ctype    en_US.UTF-8
#>  tz       America/Mazatlan
#>  date     2021-08-10
#>
#> ─ Packages ───────────────────────────────────────────────────────────────────
#>  ! package     * version  date       lib source
#>    AsioHeaders   1.16.1-1 2020-07-07 [1] RSPM (R 4.1.0)
#>    askpass       1.1      2019-01-13 [1] CRAN (R 4.1.0)
#>    assertthat    0.2.1    2019-03-21 [1] CRAN (R 4.1.0)
#>    backports     1.2.1    2020-12-09 [1] CRAN (R 4.1.0)
#>    cachem        1.0.5    2021-05-15 [1] CRAN (R 4.1.0)
#>    callr         3.7.0    2021-04-20 [1] CRAN (R 4.1.0)
#>    cli           3.0.1    2021-07-17 [1] CRAN (R 4.1.0)
#>    commonmark    1.7      2018-12-01 [1] CRAN (R 4.1.0)
#>    crayon        1.4.1    2021-02-08 [1] CRAN (R 4.1.0)
#>    credentials   1.3.1    2021-07-25 [1] RSPM (R 4.1.0)
#>    DBI           1.1.1    2021-01-15 [1] CRAN (R 4.1.0)
#>    desc          1.3.0    2021-03-05 [1] CRAN (R 4.1.0)
#>    devtools      2.4.2    2021-06-07 [1] CRAN (R 4.1.0)
#>    digest        0.6.27   2020-10-24 [1] CRAN (R 4.1.0)
#>    dplyr         1.0.7    2021-06-18 [1] CRAN (R 4.1.0)
#>    ellipsis      0.3.2    2021-04-29 [1] CRAN (R 4.1.0)
#>    evaluate      0.14     2019-05-28 [1] CRAN (R 4.1.0)
#>  P excluder    * 0.2.2    2021-08-07 [?] local
#>    fansi         0.5.0    2021-05-25 [1] CRAN (R 4.1.0)
#>    fastmap       1.1.0    2021-01-25 [1] CRAN (R 4.1.0)
#>    fs            1.5.0    2020-07-31 [1] RSPM (R 4.1.0)
#>    generics      0.1.0    2020-10-31 [1] CRAN (R 4.1.0)
#>    gert          1.3.1    2021-06-23 [1] CRAN (R 4.1.0)
#>    glue          1.4.2    2020-08-27 [1] CRAN (R 4.1.0)
#>    highr         0.9      2021-04-16 [1] CRAN (R 4.1.0)
#>    hms           1.1.0    2021-05-17 [1] CRAN (R 4.1.0)
#>    htmltools     0.5.1.1  2021-01-22 [1] CRAN (R 4.1.0)
#>    iptools       0.6.1    2018-12-09 [1] RSPM (R 4.1.0)
#>    janitor       2.1.0    2021-01-05 [1] CRAN (R 4.1.0)
#>    knitr         1.33     2021-04-24 [1] CRAN (R 4.1.0)
#>    lifecycle     1.0.0    2021-02-15 [1] CRAN (R 4.1.0)
#>    lubridate     1.7.10   2021-02-26 [1] CRAN (R 4.1.0)
#>    magrittr      2.0.1    2020-11-17 [1] CRAN (R 4.1.0)
#>    maps          3.3.0    2018-04-03 [1] RSPM (R 4.0.3)
#>    memoise       2.0.0    2021-01-26 [1] CRAN (R 4.1.0)
#>    openssl       1.4.4    2021-04-30 [1] CRAN (R 4.1.0)
#>    pillar        1.6.2    2021-07-29 [1] CRAN (R 4.1.0)
#>    pkgbuild      1.2.0    2020-12-15 [1] CRAN (R 4.1.0)
#>    pkgconfig     2.0.3    2019-09-22 [1] CRAN (R 4.1.0)
#>    pkgload       1.2.1    2021-04-06 [1] CRAN (R 4.1.0)
#>    prettyunits   1.1.1    2020-01-24 [1] CRAN (R 4.1.0)
#>    processx      3.5.2    2021-04-30 [1] CRAN (R 4.1.0)
#>    ps            1.6.0    2021-02-28 [1] CRAN (R 4.1.0)
#>    purrr         0.3.4    2020-04-17 [1] CRAN (R 4.1.0)
#>    R6            2.5.0    2020-10-28 [1] CRAN (R 4.1.0)
#>    rcmdcheck     1.3.3    2019-05-07 [1] CRAN (R 4.1.0)
#>    Rcpp          1.0.7    2021-07-07 [1] CRAN (R 4.1.0)
#>    readr         2.0.1    2021-08-10 [1] CRAN (R 4.1.0)
#>    remotes       2.4.0    2021-06-02 [1] CRAN (R 4.1.0)
#>    reprex        2.0.1    2021-08-05 [1] RSPM (R 4.1.0)
#>    rlang         0.4.11   2021-04-30 [1] CRAN (R 4.1.0)
#>    rmarkdown     2.10     2021-08-06 [1] CRAN (R 4.1.0)
#>    roxygen2      7.1.1    2020-06-27 [1] CRAN (R 4.1.0)
#>    rprojroot     2.0.2    2020-11-15 [1] CRAN (R 4.1.0)
#>    rstudioapi    0.13     2020-11-12 [1] CRAN (R 4.1.0)
#>    sessioninfo   1.1.1    2018-11-05 [1] CRAN (R 4.1.0)
#>    snakecase     0.11.0   2019-05-25 [1] CRAN (R 4.1.0)
#>    stringi       1.7.3    2021-07-16 [1] CRAN (R 4.1.0)
#>    stringr       1.4.0    2019-02-10 [1] CRAN (R 4.1.0)
#>    styler        1.5.1    2021-07-13 [1] CRAN (R 4.1.0)
#>    sys           3.4      2020-07-23 [1] CRAN (R 4.1.0)
#>    testthat    * 3.0.4    2021-07-01 [1] CRAN (R 4.1.0)
#>    tibble        3.1.3    2021-07-23 [1] RSPM (R 4.1.0)
#>    tidyr         1.1.3    2021-03-03 [1] CRAN (R 4.1.0)
#>    tidyselect    1.1.1    2021-04-30 [1] CRAN (R 4.1.0)
#>    triebeard     0.3.0    2016-08-04 [1] RSPM (R 4.1.0)
#>    tzdb          0.1.2    2021-07-20 [1] RSPM (R 4.1.0)
#>    usethis       2.0.1    2021-02-10 [1] CRAN (R 4.1.0)
#>    utf8          1.2.2    2021-07-24 [1] RSPM (R 4.1.0)
#>    vctrs         0.3.8    2021-04-29 [1] CRAN (R 4.1.0)
#>    withr         2.4.2    2021-04-18 [1] CRAN (R 4.1.0)
#>    xfun          0.25     2021-08-06 [1] CRAN (R 4.1.0)
#>    xml2          1.3.2    2020-04-23 [1] CRAN (R 4.1.0)
#>    xopen         1.0.0    2018-09-17 [1] CRAN (R 4.1.0)
#>    yaml          2.2.1    2020-02-01 [1] CRAN (R 4.1.0)
#>
#> [1] /home/mauro/R/x86_64-pc-linux-gnu-library/4.1
#> [2] /usr/local/lib/R/site-library
#> [3] /usr/lib/R/site-library
#> [4] /usr/lib/R/library
#>
#>  P ── Loaded and on-disk path mismatch.
```

-

* Inside `if()` I suggest you use `identical()` not `==` because `if()` only takes objects of length 1. `identical()` always returns an object of length 1, but the output of `==` can be longer. If you do use `==` inside `if()` then be sure to first assert the input is indeed of length 1 (e.g. with `stopifnot()`).

* When styling your code I suggest you stick to lines of 80 characters or less. The goal is to be friendly with the reviewers, who may adjust their screen to show the typical 80 characters. Longer lines force them to scroll right, which sounds like a minor issue but after a while can get annoying. The image below shows how the first line of code overflows my 80 characters marker. Under that I have the same code styled in different ways -- all fit in my screen.

* In [line 55](https://github.com/JeffreyRStevens/excluder/blob/d99bc23850c3f6f7ae3001cf0b94324b9e0a827b/R/collapse_exclusions.R#L55) I think you need to quote "exclusions". In [`?tidyr::unite()`](https://tidyr.tidyverse.org/reference/unite.html) I see the argument `col` is _The name of the new column, as a string or symbol._ I thinks symbols (e.g. `exclusions`) are allowed for backward compatibility and I believe the tidyverse now recommnds reserving symbols for existing columns and strings (e.g. "exclusions") for new columns. Note the [examples section of `unite()`'s helpflie ](https://tidyr.tidyverse.org/reference/unite.html#examples) uses strings ("z" and "xy") -- not symbols (`z` or `xy`). I tried "exclusions" and successfully removed the NOTE. [Line 60](https://github.com/JeffreyRStevens/excluder/blob/d99bc23850c3f6f7ae3001cf0b94324b9e0a827b/R/collapse_exclusions.R#L60) looks good to me.

Thanks for your work and suggestions for reviewers.
