### 2021-08-12 #455 @JeffreyRStevens
Many thanks for the clarifications and help, @maurolepore.
* I believe that I fixed the R CMD note about qualtRics and rgeolocate. I was referencing their documentation, and if you don't have those packages installed, the documentation cannot be linked. So I changed the references to external URLs rather than internal documentation.
* I have now changed all `==` operators to `identical()` in `if()` statements that use user inputted arguments. I have added `stopifnot()` statements when other operators (e.g., `%in%`) are used in `if()` statements with user inputted arguments.
* I greatly reduced the number of lines over 80 characters but did not eliminate them all (e.g., when breaking a line would cause a message to print on two lines or would make readability much worse).
* Thank you for this! The quotes fixed it.
