### 2021-08-30 #455 @maurolepore
@JeffreyRStevens, I'm thrilled that @juliasilge and @jmobrien accepted to review the excluder package.

Note `@ropensci-review-bot` set the due date following rOpenSci's guidelines to 2021-09-20. However, these are difficult times and I would like to allow reviewers 1 extra week. Feel free to submit your review whenever ready but if you don't I'll touch base on 2021-09-27.

I look forward to working with you all.
