### 2021-09-26 #455 @juliasilge
Congratulations to the author on this useful package for folks handling Qualtrics surveys. It will be convenient to have all these common checks in a consistent set of functions, and the messaging in the console is very nice. 🙌

- [x] **Examples** (that run successfully locally) for all exported functions

On documentation, I find some of the way the documentation works a bit confusing, especially when things go wrong for users. For example, if I have a dataset with no IP addresses or location information and I try to run `exclude_duplicates()`, I get this error:

```
Error in check_duplicates(x, quiet = TRUE, ...) :
The column specifying location ('location_col') was not found.
```

However, when I look at the documentation for the function I used, I don't see anything about `location_col` and it's not clear which link I should click to find the relevant info. Some options might be to change the documentation for the dots or write a more clear error message.

---

🎯 The function name `collapse_exclusions()` reminds me a bit too much of `dplyr::collapse()` when what it does is much more like `tidyr::unite()`. What about a name like `unite_exclusions()`? Along those lines, did you consider names like `filter_*()` instead of `exclude_*()`? A benefit to using names that align with other names in the ecosystem is that these functions then slot right in to people's mental model of what they are doing.

🎯 None of my real surveys have IP addresses so I was not able to check the iptools integration beyond the example survey data included in the package.

🎯 In the future, if you wanted to increase the polish of the console messages (once the nicest features of this package that I think will draw folks to use it), you might check out using [cli](https://cli.r-lib.org/).
