### 2021-09-28 #455 @jmobrien
- [X] **Function Documentation:** for all exported functions (**but see below**)

- [X] **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines (**but see below**).

---

#### Overall
I quite like this package.  It fills a niche for Qualtrics data cleaning that is difficult to automate and/or make auditable/reproducible, given how Qualtrics designed server-side data filters.

As the author notes, in human participants research it's also important to not unduly exclude data, and the manual review this helps facilitate is supportive of that ethical imperative.  So, generally, I think, the paradigm of `excluder` makes a lot of sense--a simple interface for early-stage review and removal of problem cases.

The package seemed to work as expected.  None of the data I had available (apparently) had any non-US participants, so I also was not able to test IP-checking as thoroughly as might be warranted, but it seems like a sound implementation.

#### A few notes

A few things that might warrant either review or improvement. Nothing terribly serious or strictly essential for fixing right now--though with one issue (centering everything around the `check_*()` functions), I do wonder  that it might make growing/maintaining this package harder in the long run.

* Style & dependencies
  *  seem to be a bit inconsistent
  * mostly choices to use tidyverse vs. base functions, e.g.:
    * `dplyr::bind_cols()` is mixed with `rbind()`.
    * `dplyr::pull()` is regularly used for indexing that could be handled by `[[]]`.
  * Nothing seems strictly wrong, but the motivation for each choice wasn't always clear

* some function calls reference function re-exports rather than the originals
  * e.g., `dplyr::all_of()` instead of `tidyselect::all_of()`.
  * Not currently clear which approach is preferable, but thought it worth mentioning.

* Interaction b/w `check`, `mark`, and `exclude` verbs internally & in workflow:
  * Again, I like the verbs as an overarching approach.
  * However, considering [Mozilla guidelines](https://mozillascience.github.io/codeReview/review.html) on simplicity/efficiency, I do wonder about everything depending on the `check_*()` functions
  * Outlining here:
    * `mark_*()` functions pass everything to `check_*()`,
    extracts a specified ID variable from that, adds a marker text variable,
    then merges this back with `left_join()`
    * `exclude_*()` functions pass everything to `check_*()`
    then uses `anti_join()` on just the ID variable
  * I think this arrangement might unnecessarily complicate some tasks.  For example:
    * Output from `mark_*()` functions end up with `NA` in non-marked rows of `exclusion_` variables. This  makes manual filtering on these columns more difficult.
    * `mark_duration()` can't easily mark the difference b/w a too-short and a
    too-long response, which might be a useful thing to know.
    * `mark_*()` and `exclude_*()` use ellipses for
    arguments to be passed to `check_*()`, meaning you have to look up
    `check_*()` to get comprehensive help on how `mark_*()` and `exclude_*()`
    operate (this specific issue could probably be fixed with a few explicit
    arguments in each function, even if everything else is left the same)
  *  [*personal opinion here*]-- could it be simpler & more useful if
  package functionality built on the `mark_*()` functions, rather than `check_*()`?
      * `mark_*()` functions could cover all reporting.  This could even enable some
    time-saving advantages from `summarize()`-like reporting, (e.g., could see at a
    glance if all short-duration cases were those also out-of-US).
      * `exclusion_*` variables from `mark` could use booleans, 1/0, (or even just ""),
    for easier support of manual filtering if desired
      * `check_*()` and `exclude_*()` could be simplified:
          * Under the current design (3 functions for each characteristic),
      `check` and `exclude` could just call `mark` internally plus something
      like `dplyr::filter()`.
          * Potentially, `check` and `exclude` could even be reduced to just 2
      general-purpose that keep/exclude based on any added `exclusion_` variables
      * Efficiency might be improved by avoiding duplicating passes on the data--
    maybe not noticeable enough to matter for smaller use cases
    (such as those more typical for we psychologists), potentially useful for
    potential users who work regularly with larger-scale datasets.

* Few small other things:
  * use of & instead of && in if statements that should only field one comparison
  * descriptions of `check_ip()`, `exclude_ip()`, and `mark_ip()` are backwards--each
  retains/drops cases from **outside** the specified country, not from **inside** it as indicated in the documentation
  * the rOpenSci packaging guidelines recommends an `[object]_[verb]` structure,
  while this uses the reverse.  Personally, I think current naming is fine, though.
  * Packaging guidelines generally recommend not having a separate file for each function, but having at least a bit of conceptual organization.  The names help here, but it did slow down my navigating the package.
