### 2021-09-28 #455 @maurolepore
@jmobrien thanks a lot for your review.

@JeffreyRStevens, you can now go ahead and address reviewer's comments. rOpenSci's guidelines seem to not limit how long you should take, but I think it's best to respond ASAP -- while the reviewers keep the context in their heads.

Please let me know if you have questions.
