### 2021-09-28 #455 @JeffreyRStevens
Yes, many thanks @juliasilge and @jmobrien for your helpful reviews! I've done a quick skim and will review in-depth very soon. I may have a few questions before I get too far into revisions.
