### 2021-10-03 #455 @JeffreyRStevens
@juliasilge, many thanks again for your helpful comments. Here are some responses/questions for your points:
* Confusing error/documentation for `exclude_duplicates()`: I think I see the issue here, and I believe that addressing some of @jmobrien's concerns about the functions will address this. I'll return to this comment after I address his concerns.
* Rename `collapse_exclusions()`: This was a great idea. I have now renamed this to `unite_exclusions()` and marked `collapse_exclusions()` as deprecated using the `{lifecycle}` package. I will push this change after I resolve the next concern.
* Rename `exclude_*()` functions to `filter_*()`: I really like this idea, too. But one question that I had is whether I should rather use `filter_*()` to replace the `check_*()` functions since the tidyverse `filter()` function *retains* the specific rows rather than *removes* them. In this sense, it seems to be more equivalent to `check_*()` rather than `exclude_*()`, in that `check_*()` retains the rows of data that are being checked. Or maybe I'm thinking of it backwards. What are your thoughts on this?
* Thank you so much for pointing me to `{cli}`. I hadn't seen that before, and it looks like a very useful package for customizing messages. I'll definitely be figuring out ways to integrate that into the console messages.

I'll have some questions for @jmobrien in the coming days.
