### 2021-10-04 #455 @juliasilge
I was really happy to see @jmobrien's thorough review and I agree that working through some of his suggestions will address the problems with the documentation and error messages.

I'm not sure what we really mean by a difference between "retain" vs. "throw away" (just reverse the condition, which in the case of, say, IP addresses in/out of the US doesn't seem like a super one-way thing) but I think I see your point. I guess my opinion is that creating three new verbs for operations that are arguably the same as things many people already use makes this toolkit less easy to use than if they were given names that allude to the ways in which they are the same as other operations. Another option may be to use `mark_` + `filter_` + `exclude_`?
