### 2021-10-04 #455 @jmobrien
I also will say that I generally agreed with all that @juliasilge said in her review and thought it was on point.  Where I didn't echo her sentiments directly, it was mostly for brevity because I was the second to submit and my review was already verbose.  I especially agree that the console reporting is a big sell here.

Here, @JeffreyRStevens, I'm understanding you as saying `exclude_duration([longer than 30 mins])` would remove all >30 minute cases, while `filter([duration > 30 mins])` would keep those same cases and discard others.  So, at least at default, the behavior of `filter` is more conceptually similar to `check` than `exclude`.  That right?  If so, I see where you're coming from also.

From a more data-centric perspective on the potential userbase, I agree w/where Julia's coming from.  The package does things very close to core tidyverse operations for data wrangling. There's something to be said for capitalizing on what those names invoke.

With my social scientist had on, though, I feel much more positively about the existing names.  This whole package really just seems like a batch of convenience functions that abstract lower-level data wrangling operations into key *tasks* for ethically managing human-participants data.  As tasks, I think the names make sense: manually "check" problems first, then "exclude" them prior to analysis.

I don't have a strong opinion about which way naming should go.  Though FWIW, I suppose similar ideas are embedded in my speculation whether functionality should center on `mark`.  With that workflow, the `check` and `exclude` verbs could keep their core purpose as convenience functions called downstream of `mark` (give me all the marked stuff for review vs. get rid of it).  But if you find yourself needing something more complex, you can just do a standard call to `filter` instead.

(Or, if you're feeling really ambitious, maybe you could make everyone happy by adding something like `filter_exclusions()` that could emulate the typical filter syntax for more complex cases where the convenience functions fall short?  So, e.g., `filter_exclusions(!ip, duration)` could mean "give me all the cases that were NOT marked as out of the acceptable IP address space, but which were still too long", etc.  Could maybe also work around whether `unite_exclusions()` has been called yet or not.)
