### 2021-10-04 #455 @juliasilge
I don't know that I'd recommend adding another set of functions to maintain here. And for the record, my opinion on the naming is not super strong either; I'm raising this mostly because this is something I've run into before and in the past I have always regretted when I have not used what already exists in terms of vocabulary, norms, idioms, function names, etc. and then later realized I could have.
