### 2021-10-04 #455 @JeffreyRStevens
Thanks you two for your thoughts on this. Yes, my thought was that `filter_`, `mark_`, and `exclude_` would more closely match the way `filter()` works (that is, it retains rows satisfying the condition). But I also share @jmobrien's point that `check_` might make more sense from a general user perspective even though `filter_` makes sense from a tidyverse user/developer perspective. I'll grapple with this in the coming days.
I really like the idea of centering the functions around `mark_` rather than `check_`. @jmobrien, just to clarify, you were also suggesting removing all of the `check_` and `exclude_` functions and instead only have two functions: `check` and `exclude`, correct? So basically, you would always run the `mark_` functions to find the exclusions, then pipe that to `check` to view/store the exclusions or pipe to `exclude` to remove the exclusions. Is this what you were suggesting? This sounds much easier to maintain and probably to use as well.
