### 2021-10-05 #455 @jmobrien
That was one option, yes.  I guess in that simplifying approach `check` and `exclude` would need to recognize what parts that `mark` functions had looked at previously.  You could set that up a number of different ways.

You could also preserve the existing `check`/`exclude` function sets if you prefer a more standalone design where each verb can operate on fresh data directly.  Even then, as outlined above I think there are several benefits to having `check` and `exclude` call `mark` internally, rather than `mark` and `exclude` calling `check`.

There's certainly some draw to option 1.  But both options have unique advantages, and I don't want to dictate your design choices.

Option 1's a more major change, though, so if you're leaning that way I might get others' opinions first.
