### 2021-10-09 #455 @JeffreyRStevens
I'm close to finishing up revisions. I have a couple more changes to make and had two questions for @jmobrien.
1. When you referred to avoiding multiple passes on the data, did you mean within a function or are you concerned about running multiple `mark_*()` functions?
2. When you referred to using ellipses to pass arguments to (now) `check_*()` and `exclude+*()` functions, you suggested explicitly including some arguments. Can you point me to how to do this? My efforts to do this have not worked out.
