### 2021-10-13 #455 @JeffreyRStevens
OK, I think that I have addressed all of the concerns. Please let me know if I missed something or you have additional questions. Note that I have increased the version to 0.3.0 since I have made major changes to the code and deprecated a function.

## Responses to @juliasilge.
* Confusing error/documentation for `exclude_duplicates()`: This has been addressed by explicitly including arguments from `mark_*()`. [Commit 1b74c4ce](https://github.com/JeffreyRStevens/excluder/commit/1b74c4ce1d982d2af816cdc2b37bbdf6538887b8)
* Rename `collapse_exclusions()`: This was a great idea. I have now renamed this to `unite_exclusions()` and marked `collapse_exclusions()` as deprecated using the `{lifecycle}` package. [Commit 68aae92b](https://github.com/JeffreyRStevens/excluder/commit/68aae92beade3b14d8018fd55f205f42e8389c2e#diff-8c2574892063f995fdf756bce07f46c1a5193e54cd52837ed91e32008ccf41ac).
* Rename `exclude_*()` functions to `filter_*()`: I'm still torn on this, but I think that I will keep the names as originally named. In part, this is due to the confusion that we've had about whether `check` or `exclude` should be changed to `filter`. I now use filter for both types of actions. Maybe a compromise is to use the word 'filter' more in the documentation?
* Thank you so much for pointing me to `{cli}`. I hadn't seen that before, and it looks like a very useful package for customizing messages. I'll definitely be figuring out ways to integrate that into the console messages.

## Responses to @jmobrien
* Using both `dplyr::bind_cols()` and `rbind()`: I have switched all `rbind()` uses to `dplyr::bind_cols`. [Commit 89294932](https://github.com/JeffreyRStevens/excluder/commit/89294932e1c4412cc986c1b69dc359484af9b600)
* Using `dplyr::pull()` instead of `[[]]`: I have switched all `dplyr::pull()` uses to `[[]]`. [Commit ec655f9b](https://github.com/JeffreyRStevens/excluder/commit/ec655f9bdbe7d8a73655999d4a204e53a7016b40)
* Referencing function re-exports instead of original package: I have changed all references to `dplyr::all_of()` and `dplyr::any_of()` to `tidyselect::all_of()` and `tidyselect::any_of()`. Are there ways to easily find the original package for all of your functions in a script? [Commit face90d](https://github.com/JeffreyRStevens/excluder/commit/face90d864784f92fcd633af884e77339e2fb8e3)
* Move primary functionality from `check_*()` to `mark_*()` functions: I have now done this. The `check_*()` and `exclude_*()` functions now just run the `mark_*()` functions and filter exclusions in or out. [Commit 877e3b9](https://github.com/JeffreyRStevens/excluder/commit/877e3b9b33d12147950545516c29e9d1fef3e43d)
* Mark functions result in NAs when no exclusions apply: The NAs have now been replaced with "". [Commit 877e3b9](https://github.com/JeffreyRStevens/excluder/commit/877e3b9b33d12147950545516c29e9d1fef3e43d)
* `mark_duration()` does not differentiate between too long and too short: `mark_duration()` now marks quick and slow durations separately. [Commit d45e44a6](https://github.com/JeffreyRStevens/excluder/commit/d45e44a630bef06651798d09adc53710525fcd6b)
* `mark_*()` and `exclude_*()` use ellipses for arguments to be passed to `check_*()`: This has been addressed by explicitly including arguments from `mark_*()`. [Commit 1b74c4ce](https://github.com/JeffreyRStevens/excluder/commit/1b74c4ce1d982d2af816cdc2b37bbdf6538887b8)
* Create two general-purpose functions `check` and `exclude`: I've decided to hold off on this for the moment. Instead, I have offered an example in the vignette on piping multiple `mark_*()` functions and then check with a filter. [Commit 12b4ad2b](https://github.com/JeffreyRStevens/excluder/commit/12b4ad2b9622184bccc747d5999696ced8e2dc8d)
* Avoid duplicate passes on data: This may be an optimization issue to deal with later.
* Use `&&` instead of `&`: I have now replaced `&` with `&&`. [Commit b9aa1b04](https://github.com/JeffreyRStevens/excluder/commit/b9aa1b04ff09a0bc0931f7bf13797104fb76e7a7)
* The ip functions were inaccurately described in documentation: I have fixed the documentation to clarify that these functions find cases *outside* of the specified country. [Commit 3b486a20](https://github.com/JeffreyRStevens/excluder/commit/3b486a2085382898685388094ffa1ec2b617561e)
* Reverse order of function names to `[object]_[verb]`: While I saw that this was a preference for rOpenSci packages, it does not seem natural for this set of functions. `[verb]_[object]` both categorizes the verbs better and matches {tidyverse} syntax (e.g., `bind_rows()`/`bind_cols()`) as well as the {qualtRics} syntax (e.g., `fetch_survey()`).
* Organize related functions into files: I have now organized the functions by the exclusion type (e.g., `duplicates.R`). [Commit 3c88d484](https://github.com/JeffreyRStevens/excluder/commit/3c88d484d75ca9a2ebb9b3bafac699a74fed06ab)
