### 2021-10-14 #455 @maurolepore
Thanks everyone! It's been great to watch this review.

Dear reviewers @juliasilge and @jmobrien,

> At this stage, we ask that you respond as to whether the changes sufficiently address any issues raised in your review. We encourage ongoing discussion between package authors and reviewers, and you may ask editors to clarify issues in the review thread as well.
> -- https://devguide.ropensci.org/reviewerguide.html#followupreviewer
