### 2021-10-30 #455 @JeffreyRStevens
@maurolepore, what are the next steps now? FYI, I'm working on a paper for JOSS. I'm not sure if that should be part of the review here or just over at JOSS.
