### 2021-10-30 #455 @maurolepore
@JeffreyRStevens,

I'm watching your discussion with the reviewers. Do you feel the ball is no longer on your court? If so I would ask the reviewers once again whether your changes sufficiently address any issues they raised in the review. Please confirm and I'll write a new comment with a `@mention` to them.

We no longer review submissions to JOSS, but do mention the review -- I think you might get fast-tracking.
