### 2021-10-31 #455 @maurolepore
Dear reviewers,

@jmobrien, do you have an estimate about when you might be able to confirm whether the changes made are sufficient to approve this package?

@juliasilge, it seems your last concern (return visibly) has been addressed. Is there anything else you'd like to request or do you confirm the changes made are sufficient to approve this package?
