### 2021-10-31 #455 @jmobrien
Mauro, and Jeffrey, apologies--I've had a remarkable variety of unexpected
home and family issues come up over just this past week.  I'm about to go
handle what I hope is the last of them, but I'll make it a priority to
finish everything tomorrow.  Thanks for your patience.

On Sat, Oct 30, 2021 at 8:37 PM Mauro Lepore ***@***.***>
wrote:

> Dear reviewers,
>
> @jmobrien <https://github.com/jmobrien>, do you have an estimate about
> when you might be able to confirm whether the changes made are sufficient
> to approve this package?
>
> @juliasilge <https://github.com/juliasilge>, it seems your last concern
> (return visibly) has been addressed. Is there anything else you'd like to
> request or do you confirm the changes made are sufficient to approve this
> package?
>
