### 2021-11-01 #455 @jmobrien
This looks much better.  The modularization achieves a lot of the code streamlining I was hoping for, especially.

I think I'm ready to approve, except for one maybe-bug and maybe 1 or 2 suggestions:

**1. Small bug(?):**

`mark_ip()` checks for missing IPs (previews, mostly) and notes them, but doesn't actually mark those rows.  Thus, they aren't kept in `check_ip()` and they're left in with `exclude_ip()`.  If this is intentional (because `mark_preview()` exists), then fine, but otherwise this should be fixed.

**2. marking columns in `check_*()` functions**

Currently `check_()` functions drop the marking variable columns. But since `check_()*` seems intended for manual review, and the marks sometimes have useful summary notes (e.g., duration).  Would it make sense to keep them in, or at least have an argument for that?  If so, I also might move them to the first column in that case for easier viewing if a survey is long.

3. Exclusions for `exclude_*()`?

Is the philosophy for having `check_*()` functions that, following review, you might sometimes see cases you still want to keep?  If so and that occurs, it seems like `exclude_*()` stops working well.  There's no way to note which cases you want to protect (other than modifying the actual data, which presumably you don't want to do).

I'm not sure how you would want to approach an explicit feature for this, so I don't think adding one should hold up approval. Maybe something to consider for the future.  Still, if this does fit the intended use for `check_*()`, maybe a simple addition to the vignette covering how to do this with a `mark` --> [adjustments] --> `filter` workflow in lieu of `exclude` would be good for users to see.

I think all of the above should be pretty trivial to implement, and ignore them if they don't fit your goals.  Other than that, I think this is ready to go!  Great work!
