### 2021-11-02 #455 @JeffreyRStevens
@jmobrien, @juliasilge, @maurolepore, I have a question for you all.
I'm working on the issues raised above and have created a new function (in `utils.R`) called `keep_marked_column()` that is used in the `check_*()` functions. This function reads in the data, the name of the exclusion column that is created by the `mark_*()` function, and a flag for whether to keep the column in the output (as per item 2 above). It then simply removes or relocates the exclusion column based on the flag. However, I'm running into a problem with passing R CMD checks.
When I run the R CMD check on the package with the function as it is currently written (and pushed), I receive the dreaded `no visible binding for global variable` note. This seems to result from when I embrace the column name with `{{ }}` in line 31: `x %>% dplyr::select(-{{ column }})`. If I switch to an indirect call of the column name with `.data[[column]]` (as used in line 33), I receive an error that the column name is not found (e.g., when running `check_duplicates()`, I get `Error: object 'exclusion_duplicates' not found`). I can get it to work with `.data[[!!rlang::enquo(column)]]`, but then I get the check error again.
Is there a more elegant way of resolving this besides assigning all of the columns to a `globalVariables()` call?
