### 2021-11-02 #455 @jmobrien
@JeffreyRStevens took a glance at it on your page.  Just a thought, but I'm actually wondering if simplest solution to what you're facing may be to just make the `column` argument of `keep_marked_column()` take a string instead of a bare name.  Then, just use `all_of()` inside each of the `select` and `relocate` functions?

The name being passed into `keep_marked_column()` is known and fixed for each `check` function, so I don't think you need anything more flexible.  And the name itself is really all you'll ever need out of that argument, so a string seems like it could be fully adequate?

(You could also do something similar using `deparse(substitute(column))`, or some more "tidy" equivalent, inside the function)
