### 2021-11-02 #455 @jmobrien
Working with the current approach, might this work?

I haven't tested it in a build and @JeffreyRStevens might know something better/more best-practices, but I think it's an option.

```r
keep_marked_column <- function(x, column, keep) {
  col <- rlang::ensym(column)
  if (identical(keep, FALSE)) {
    x %>% dplyr::select(!!col)
  } else {
    x %>% dplyr::relocate(!!col)
  }
}
````
