### 2021-11-02 #455 @maurolepore
@JeffreyRStevens,

My local experiment suggests that adding `.data$` to your column names avoids the R CMD check NOTE. To confirm the checks on GitHub Actions I believe you'll need to approve the workflows somehow (apparently because I'm a first-time contributor to your package).

https://github.com/JeffreyRStevens/excluder/pull/5/files

My suggestion is independent to the advice that @jmobrien shared.

--

Also, while I'm here, you may simplify the `if()` clause:

``` r
keep <- FALSE
out <- "it's true"
if (!keep) {
  out <- "it's false"
}
out
#> [1] "it's false"

keep <- TRUE
out <- "it's true"
if (!keep) {
  out <- "it's false"
}
out
#> [1] "it's true"
```

See https://speakerdeck.com/jennybc/code-smells-and-feels?slide=44
