### 2021-11-02 #455 @JeffreyRStevens
@jmobrien
1. Small bug(?):
I have now added a flag in `mark_ip()` that tracks whether to include NAs.

2. marking columns in check_*() functions
I have now added a flag in the `check*_()` functions that tracks whether to keep (and move to the first column) the exclusion column or remove it.

3. Exclusions for exclude_*()?
This is an interesting idea that I would like to think about a bit before doing anything. If you're OK with it @jmobrien, I'd like to wait to deal with this until after approval.

Is there anything else?
