### 2021-11-04 #455 @maurolepore
> I think all of the above should be pretty trivial to implement, and ignore them if they don't fit your goals. Other than that, I think this is ready to go! Great work!
-- https://github.com/ropensci/software-review/issues/455#issuecomment-956550189

The way I interpret @jmobrien's comment is that he didn't need another round of review. You have addressed all his latest comments, so congratulations @JeffreyRStevens -- excluder is approved!

And once again thanks @juliasilge and @jmobrien for your amazing work!

I'll follow up soon with some more comments and instructions.
