### 2021-11-04 #455 @maurolepore
@JeffreyRStevens I think the comment above includes everything you except two things:

1. As you intend to submit to CRAN, please see https://devdevguide.netlify.app/building.html#crangotchas

2. See the links under [Package promotion](https://devdevguide.netlify.app/editorguide.html#package-promotion).
