### 2021-11-04 #455 @JeffreyRStevens
@maurolepore, I believe I have fully transitioned the package over to rOpenSci. The one remaining issue that I can't sort out is how to change the URL link in the About section of the GitHub page (https://github.com/ropensci/excluder) to https://docs.ropensci.org/excluder/. It is currently still linking to jeffreyrstevens.github.io/excluder even though I have changed all of the links in the repo. Do you know how to change this or will it just resolve on its own at some point?
