### 2021-11-04 #455 @maelle
@JeffreyRStevens I've made you admin of your repo again, so you should be able to make changes now.

Admin rights are always lost during transfer, but we're working on a bot command for allowing package authors themselves to reclaim them. You'll experience a smoother experience if you submit a second package. :wink:
